feat(domain): define currency entities and repository contract

+++ lib/features/currency_converter/domain/entities/exchange_rate.dart
@@ -1,9 +1,18 @@
+import 'package:decimal/decimal.dart';
 import 'package:equatable/equatable.dart';
 
+/// Represents a single currency exchange rate for a given base.
+/// Holds the target currency code and the multiplier rate needed for conversion calculations.
 class ExchangeRate extends Equatable {
+  /// The currency code being converted from (e.g., 'USD').
   final String baseCurrency;
+
+  /// The currency code being converted to (e.g., 'AED').
   final String targetCurrency;
-  final double rate;
+
+  /// The multiplier applied to the base amount.
+  /// Uses Decimal instead of double to prevent floating point precision errors.
+  final Decimal rate;
 
   const ExchangeRate({
     required this.baseCurrency,
